ch #163307276 Restructured my MVC style
