--- server/models/questions.json
@@ -1,62 +1,0 @@
-[
-  {
-    "id": 1,
-    "createdOn": "2019-01-17 14:53:19",
-    "createdBy": 5,
-    "meetup": 105,
-    "title": "Fellowship Talks",
-    "body": "We shall talk about everything on fellowship and how to get started",
-    "upvotes": 5,
-    "downvotes": 3
-  },
-  {
-    "id": 2,
-    "createdOn": "2019-01-17 14:53:32",
-    "createdBy": 5,
-    "meetup": 105,
-    "title": "Fellowship Talks",
-    "body": "We shall talk about everything on fellowship and how to get started",
-    "upvotes": 4,
-    "downvotes": 1
-  },
-  {
-    "id": 3,
-    "createdOn": "2019-01-17 14:53:34",
-    "createdBy": 5,
-    "meetup": 105,
-    "title": "Fellowship Talks",
-    "body": "We shall talk about everything on fellowship and how to get started",
-    "upvotes": 0,
-    "downvotes": 0
-  },
-  {
-    "id": 4,
-    "createdOn": "2019-01-17 14:58:52",
-    "createdBy": 5,
-    "meetup": 105,
-    "title": "Fellowship Talks",
-    "body": "We shall talk about everything on fellowship and how to get started",
-    "upvotes": 0,
-    "downvotes": 0
-  },
-  {
-    "id": 5,
-    "createdOn": "2019-01-17 19:41:48",
-    "createdBy": 5,
-    "meetup": 3,
-    "title": "Test Fellowship Talks",
-    "body": "We shall talk about everything on fellowship and how to get started",
-    "upvotes": 0,
-    "downvotes": 0
-  },
-  {
-    "id": 6,
-    "createdOn": "2019-01-17 19:44:03",
-    "createdBy": 5,
-    "meetup": 3,
-    "title": "Test Fellowship Talks",
-    "body": "We shall talk about everything on fellowship and how to get started",
-    "upvotes": 0,
-    "downvotes": 0
-  }
-]
